requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
Django==5.1.3
django-crispy-forms==2.3
dnspython==2.8.0
email-validator==2.3.0
fastar==0.8.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.11
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.3
rignore==0.7.6
sentry-sdk==2.55.0
shellingham==1.5.4
SQLAlchemy==2.0.49
sqlparse==0.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
urllib3==2.3.0
websockets==16.0
